requirements.txt
keyboard>=0.13.5
pyttsx3>=2.90
sounddevice>=0.4.6
soundfile>=0.12.1
discord.py[voice]>=2.7.1
davey>=0.1.0
aiohttp>=3.13.4
requests>=2.33.0
python-dotenv>=1.0.0
gtts>=2.3.0
edge-tts>=7.2.8
pystray>=0.19.0
Pillow>=12.2.0
psycopg[binary]>=3.2.0
